Introduce `CompatibilityCheckDetails`
This commit introduces the `CompatibilityCheckDetails<T>` interface, to be used as a parameter of
attribute compatibility checks, instead of relyin on a `boolean`. This will allow us to express
more cases, like telling what to do when the consumer wants a value but the producer doesn't.

This is just a first step in a better matching API. This commit significantly increases the complexity
of the code you need to write for compatibility matches, but it's a temporary issue only until
we do something similar for `selectBestMatches`.